+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Disabled patching for CM on shaders with anisotropic_lighting and soft_lighting
 - Re-enabled CM patching on texture sets with slots in 3,6,7
 - Fixed non-ascii chars not working for paths in modorganizer.ini
+- Issues with plugins will now be printed to console instead of an unhandled exception
 
 ## [0.8.6] - 2025-01-28